+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/KtLint.kt
@@ -8,16 +8,16 @@ import com.pinterest.ktlint.core.api.FeatureInAlphaState
 import com.pinterest.ktlint.core.api.UsesEditorConfigProperties
 import com.pinterest.ktlint.core.internal.EditorConfigGenerator
 import com.pinterest.ktlint.core.internal.EditorConfigLoader
-import com.pinterest.ktlint.core.internal.EditorConfigLoader.Companion.convertToRawValues
 import com.pinterest.ktlint.core.internal.KotlinPsiFileFactoryProvider
 import com.pinterest.ktlint.core.internal.LineAndColumn
 import com.pinterest.ktlint.core.internal.SuppressionLocator
+import com.pinterest.ktlint.core.internal.SuppressionLocatorBuilder
+import com.pinterest.ktlint.core.internal.VisitorProvider
 import com.pinterest.ktlint.core.internal.buildPositionInTextLocator
-import com.pinterest.ktlint.core.internal.buildSuppressedRegionsLocator
-import com.pinterest.ktlint.core.internal.noSuppression
 import java.nio.file.FileSystems
 import java.nio.file.Path
 import java.nio.file.Paths
+import java.util.Locale
 import org.ec4j.core.model.PropertyType
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.lang.FileASTNode
@@ -29,13 +29,6 @@ import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.psi.KtFile
 
 public object KtLint {
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46.",
-        replaceWith = ReplaceWith("EDITOR_CONFIG_PROPERTIES_USER_DATA_KEY")
-    )
-    // TODO: when removing, then also remove method "EditorConfigProperties.toUserData"
-    public val EDITOR_CONFIG_USER_DATA_KEY: Key<EditorConfig> = Key<EditorConfig>("EDITOR_CONFIG")
-
     public val ANDROID_USER_DATA_KEY: Key<Boolean> = Key<Boolean>("ANDROID")
     public val FILE_PATH_USER_DATA_KEY: Key<String> = Key<String>("FILE_PATH")
     private const val FILE_PATH_PROPERTY = "file_path"
@@ -50,31 +43,6 @@ public object KtLint {
 
     @OptIn(FeatureInAlphaState::class)
     private val editorConfigGenerator = EditorConfigGenerator(editorConfigLoader)
-
-    /**
-     * @param fileName path of file to lint/format
-     * @param text Contents of file to lint/format
-     * @param ruleSets a collection of "RuleSet"s used to validate source
-     * @param userData Map of user options
-     * @param cb callback invoked for each lint error
-     * @param script true if this is a Kotlin script file
-     * @param editorConfigPath optional path of the .editorconfig file (otherwise will use working directory)
-     * @param debug True if invoked with the --debug flag
-     */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46.",
-        replaceWith = ReplaceWith("ExperimentalParams")
-    )
-    public data class Params(
-        val fileName: String? = null,
-        val text: String,
-        val ruleSets: Iterable<RuleSet>,
-        val userData: Map<String, String> = emptyMap(),
-        val cb: (e: LintError, corrected: Boolean) -> Unit,
-        val script: Boolean = false,
-        val editorConfigPath: String? = null,
-        val debug: Boolean = false
-    )
 
     /**
      * @param fileName path of file to lint/format
@@ -96,7 +64,6 @@ public object KtLint {
      * Enables some internals workarounds for Kotlin Compiler initialization.
      * Usually you don't need to use it and most probably it will be removed in one of next versions.
      */
-    @FeatureInAlphaState
     public data class ExperimentalParams(
         val fileName: String? = null,
         val text: String,
@@ -152,64 +119,18 @@ public object KtLint {
             .toSet()
     }
 
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46.",
-        replaceWith = ReplaceWith("ExperimentalParams")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    private fun toExperimentalParams(params: Params): ExperimentalParams =
-        ExperimentalParams(
-            fileName = params.fileName,
-            text = params.text,
-            ruleSets = params.ruleSets,
-            userData = params.userData,
-            cb = params.cb,
-            script = params.script,
-            editorConfigPath = params.editorConfigPath,
-            debug = params.debug
-        )
-
     /**
      * Check source for lint errors.
      *
      * @throws ParseException if text is not a valid Kotlin code
      * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
      */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Move '.editorconfig' properties from 'Params.userData' to 'ExperimentalParam.editorConfigOverride'.",
-        replaceWith = ReplaceWith("lint(toExperimentalParams(params))")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    public fun lint(params: Params) =
-        lint(toExperimentalParams(params))
-
-    @OptIn(FeatureInAlphaState::class)
-    public fun lint(experimentalParams: ExperimentalParams) {
-        lint(
-            experimentalParams,
-            VisitorProvider(
-                ruleSets = experimentalParams.ruleSets,
-                debug = experimentalParams.debug
-            )
-        )
-    }
-
-    /**
-     * Check source for lint errors.
-     *
-     * @throws ParseException if text is not a valid Kotlin code
-     * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
-     */
-    @FeatureInAlphaState
-    public fun lint(
-        params: ExperimentalParams,
-        visitorProvider: VisitorProvider
-    ) {
+    public fun lint(params: ExperimentalParams) {
         val psiFileFactory = kotlinPsiFileFactoryProvider.getKotlinPsiFileFactory(params.isInvokedFromCli)
         val preparedCode = prepareCodeForLinting(psiFileFactory, params)
         val errors = mutableListOf<LintError>()
 
-        visitorProvider
+        VisitorProvider(params.ruleSets, params.debug)
             .visitor(params.ruleSets, preparedCode.rootNode)
             .invoke { node, rule, fqRuleId ->
                 // fixme: enforcing suppression based on node.startOffset is wrong
@@ -273,18 +194,20 @@ public object KtLint {
             params.debug
         )
 
-        val mergedUserData = editorConfigProperties.toUserData() + params.userData
-            .run {
-                if (!params.isStdIn) {
-                    plus(FILE_PATH_PROPERTY to params.normalizedFilePath.toString())
-                } else {
-                    this
+        val mergedUserData =
+            params
+                .userData
+                .run {
+                    if (!params.isStdIn) {
+                        plus(FILE_PATH_PROPERTY to params.normalizedFilePath.toString())
+                    } else {
+                        this
+                    }
                 }
-            }
 
         injectUserData(rootNode, editorConfigProperties, mergedUserData)
 
-        val suppressedRegionLocator = buildSuppressedRegionsLocator(rootNode)
+        val suppressedRegionLocator = SuppressionLocatorBuilder.buildSuppressedRegionsLocator(rootNode)
 
         return PreparedCode(
             rootNode,
@@ -292,9 +215,6 @@ public object KtLint {
             suppressedRegionLocator
         )
     }
-
-    @Deprecated("Remove in Ktlint 0.46 when deleting EDITOR_CONFIG_USER_DATA_KEY.")
-    private fun EditorConfigProperties.toUserData() = convertToRawValues()
 
     @Deprecated(
         message = "Should not be a part of public api. Will be removed in future release.",
@@ -316,7 +236,6 @@ public object KtLint {
     ) {
         val android = userData.isAndroidCodeStyle
         node.putUserData(FILE_PATH_USER_DATA_KEY, userData[FILE_PATH_PROPERTY])
-        node.putUserData(EDITOR_CONFIG_USER_DATA_KEY, EditorConfig.fromMap(userData - "android" - "file_path"))
         node.putUserData(EDITOR_CONFIG_PROPERTIES_USER_DATA_KEY, editorConfigProperties)
         node.putUserData(ANDROID_USER_DATA_KEY, android)
         node.putUserData(
@@ -325,76 +244,23 @@ public object KtLint {
         )
     }
 
-    @Deprecated(
-        message = "Should not be a part of public api. Will be removed in future release.",
-        level = DeprecationLevel.WARNING
-    )
-    public fun calculateLineColByOffset(
-        text: String
-    ): (offset: Int) -> Pair<Int, Int> {
-        return buildPositionInTextLocator(text)
-    }
-
     /**
      * Fix style violations.
      *
      * @throws ParseException if text is not a valid Kotlin code
      * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
      */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Move '.editorconfig' properties from 'Params.userData' to 'ExperimentalParam.editorConfigOverride'.",
-        replaceWith = ReplaceWith("lint(toExperimentalParams(params))")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    public fun format(params: Params): String =
-        format(toExperimentalParams(params))
-
-    @OptIn(FeatureInAlphaState::class)
-    public fun format(experimentalParams: ExperimentalParams): String {
-        return format(
-            experimentalParams,
-            experimentalParams.ruleSets,
-            VisitorProvider(
-                ruleSets = experimentalParams.ruleSets,
-                debug = experimentalParams.debug
-            )
-        )
-    }
-
-    /**
-     * Fix style violations.
-     *
-     * @throws ParseException if text is not a valid Kotlin code
-     * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
-     */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Overrides of '.editorconfig' properties no longer should be " +
-            "passed via the 'Params.userData' but via the 'ExperimentalParam.editorConfigOverride' parameter. The " +
-            "ruleSets have to be provided via the 'ExperimentalParams.ruleSets' parameter.",
-        replaceWith = ReplaceWith("format(params, visitorProvider)")
-    )
-    @FeatureInAlphaState
-    public fun format(
-        params: ExperimentalParams,
-        ruleSets: Iterable<RuleSet>,
-        visitorProvider: VisitorProvider
-    ): String =
-        format(
-            params.copy(ruleSets = ruleSets),
-            visitorProvider
-        )
-
-    @FeatureInAlphaState
-    public fun format(
-        params: ExperimentalParams,
-        visitorProvider: VisitorProvider
-    ): String {
+    public fun format(params: ExperimentalParams): String {
         val hasUTF8BOM = params.text.startsWith(UTF8_BOM)
         val psiFileFactory = kotlinPsiFileFactoryProvider.getKotlinPsiFileFactory(params.isInvokedFromCli)
         val preparedCode = prepareCodeForLinting(psiFileFactory, params)
 
         var tripped = false
         var mutated = false
+        val visitorProvider = VisitorProvider(
+            ruleSets = params.ruleSets,
+            debug = params.debug
+        )
         visitorProvider
             .visitor(
                 params.ruleSets,
@@ -411,8 +277,8 @@ public object KtLint {
                             tripped = true
                             if (canBeAutoCorrected) {
                                 mutated = true
-                                if (preparedCode.suppressedRegionLocator !== noSuppression) {
-                                    preparedCode.suppressedRegionLocator = buildSuppressedRegionsLocator(
+                                if (preparedCode.suppressedRegionLocator !== SuppressionLocatorBuilder.noSuppression) {
+                                    preparedCode.suppressedRegionLocator = SuppressionLocatorBuilder.buildSuppressedRegionsLocator(
                                         preparedCode.rootNode
                                     )
                                 }
@@ -527,7 +393,7 @@ public object KtLint {
     }
 
     private fun determineLineSeparator(fileContent: String, userData: Map<String, String>): String {
-        val eol = userData["end_of_line"]?.trim()?.toLowerCase()
+        val eol = userData["end_of_line"]?.trim()?.lowercase(Locale.getDefault())
         return when {
             eol == "native" -> System.lineSeparator()
             eol == "crlf" || eol != "lf" && fileContent.lastIndexOf('\r') != -1 -> "\r\n"