+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/NoConsecutiveBlankLinesRuleTest.kt
@@ -3,8 +3,6 @@ package com.pinterest.ktlint.ruleset.standard
 import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThat
 import com.pinterest.ktlint.test.LintViolation
 import com.pinterest.ktlint.test.MULTILINE_STRING_QUOTE
-import com.pinterest.ktlint.test.lint
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
@@ -160,18 +158,17 @@ class NoConsecutiveBlankLinesRuleTest {
 
     @Test
     fun `should not raise NPE on linting Kotlin script file`() {
-        assertThat(
-            NoConsecutiveBlankLinesRule().lint(
-                """
-                import java.net.URI
+        val code =
+            """
+            import java.net.URI
 
-                plugins {
-                    `java-library`
-                }
-                """.trimIndent(),
-                script = true
-            )
-        ).isEmpty()
+            plugins {
+                `java-library`
+            }
+            """.trimIndent()
+        noConsecutiveBlankLinesRuleAssertThat(code)
+            .asKotlinScript()
+            .hasNoLintViolations()
     }
 
     @Test