+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/PrintASTSubCommand.kt
@@ -1,14 +1,13 @@
 package com.pinterest.ktlint.internal
 
-import com.pinterest.ktlint.KtlintCommandLine
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.ParseException
 import com.pinterest.ktlint.core.RuleSet
-import com.pinterest.ktlint.core.VisitorProvider
 import com.pinterest.ktlint.core.initKtLintKLogger
 import com.pinterest.ruleset.test.DumpASTRule
 import java.io.File
 import java.nio.file.FileSystems
+import java.util.Locale
 import mu.KotlinLogging
 import picocli.CommandLine
 
@@ -50,24 +49,19 @@ internal class PrintASTSubCommand : Runnable {
     override fun run() {
         commandSpec.commandLine().printHelpOrVersionUsage()
 
-        val visitorProvider = VisitorProvider(
-            ruleSets = astRuleSet,
-            debug = ktlintCommand.debug
-        )
         if (stdin) {
-            printAST(visitorProvider, KtLint.STDIN_FILE, String(System.`in`.readBytes()))
+            printAST(KtLint.STDIN_FILE, String(System.`in`.readBytes()))
         } else {
             FileSystems.getDefault()
                 .fileSequence(patterns)
                 .map { it.toFile() }
                 .forEach {
-                    printAST(visitorProvider, it.path, it.readText())
+                    printAST(it.path, it.readText())
                 }
         }
     }
 
     private fun printAST(
-        visitorProvider: VisitorProvider,
         fileName: String,
         fileContent: String
     ) {
@@ -84,12 +78,15 @@ internal class PrintASTSubCommand : Runnable {
                 fileName = fileName,
                 fileContents = fileContent,
                 ruleSets = astRuleSet,
-                visitorProvider = visitorProvider,
                 debug = ktlintCommand.debug
             )
         } catch (e: Exception) {
             if (e is ParseException) {
-                throw ParseException(e.line, e.col, "Not a valid Kotlin file (${e.message?.toLowerCase()})")
+                throw ParseException(
+                    e.line,
+                    e.col,
+                    "Not a valid Kotlin file (${e.message?.lowercase(Locale.getDefault())})"
+                )
             }
             throw e
         }